refactor: 重新组织代码结构
