+++ .vscode/settings.json
@@ -9,5 +9,6 @@
     "Hypergrayph",
     "PRRH",
     "Skland"
-  ]
+  ],
+  "typescript.tsdk": "node_modules/typescript/lib"
 }
