--- src/composables/storages/index.ts
@@ -1,4 +1,0 @@
-export * from './account'
-export * from './device-id'
-export * from './info'
-export * from './preference'
