--- src/composables/storages/info.ts
@@ -1,5 +1,0 @@
-import type { BindingInfo } from '~/types'
-
-export function useArknightAccountsInfo() {
-  return useWxtStorageAsync<Record<string, BindingInfo>>('PRRH:ARKNIGHT_ACCOUNTS_INFO', {})
-}
