+++ src/entrypoints/background/api/hypergrayph.ts
