+++ src/entrypoints/background/api/index.ts
