+++ src/entrypoints/background/utils/crypto.ts
