+++ src/entrypoints/background/utils/device-id.ts
@@ -1,5 +1,6 @@
+import type { DESRule } from 'background/utils/crypto'
+import { encryptAES, encryptObjectByDESRules, encryptRSA, md5 } from 'background/utils/crypto'
 import { format } from 'date-fns'
-import { encryptAES, encryptObjectByDESRules, md5 } from '~/utils/crypto'
 
 const SKLAND_SM_CONFIG = {
   organization: 'UWXspnCCJN4sfYlNfqps',
