+++ src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './logger'
-export * from './signature'
 export * from './time'
